Remove the old Coinmarketcap scrapers
