+++ config/scrapers_config.exs
@@ -9,21 +9,9 @@ config :sanbase, Sanbase.ExternalServices.Coinmarketcap,
   update_interval: 5 * 1000 * 60,
   api_key: {:system, "COINMARKETCAP_API_KEY", ""},
   api_url: "https://sandbox-api.coinmarketcap.com/",
-  sync_enabled: {:system, "COINMARKETCAP_PRICES_ENABLED", false}
-
-# TODO: Change after switching over to only this cmc
-config :sanbase, Sanbase.ExternalServices.Coinmarketcap2,
-  # 5 minutes
-  update_interval: 5 * 1000 * 60,
   sync_enabled: {:system, "COINMARKETCAP_SCRAPER_ENABLED", false}
 
 config :sanbase, Sanbase.ExternalServices.Coinmarketcap.TickerFetcher,
-  update_interval: 5 * 1000 * 60,
-  sync_enabled: {:system, "COINMARKETCAP_TICKERS_ENABLED", false},
-  top_projects_to_follow: {:system, "TOP_PROJECTS_TO_FOLLOW", "25"}
-
-# TODO: Change after switching over to only this cmc
-config :sanbase, Sanbase.ExternalServices.Coinmarketcap.TickerFetcher2,
   update_interval: {:system, "COINMARKETCAP_API_CALL_INTERVAL", "300"},
   projects_number: {:system, "COINMARKETCAP_API_PROJECTS_NUMBER", "2500"},
   sync_enabled: {:system, "COINMARKETCAP_TICKER_FETCHER_ENABLED", false},
