+++ lib/sanbase/external_services/coinmarketcap/logo_fetcher.ex
