+++ lib/sanbase/external_services/coinmarketcap/ticker_fetcher.ex
@@ -1,20 +1,21 @@
 defmodule Sanbase.ExternalServices.Coinmarketcap.TickerFetcher do
-  # # Syncronize ticker data from coinmarketcap.com
-  #
-  # A GenServer, which updates the data from coinmarketcap on a regular basis.
-  # On regular intervals it will fetch the data from coinmarketcap and insert it
-  # into a local DB
+  @moduledoc ~s"""
+    A GenServer, which updates the data from coinmarketcap on a regular basis.
+
+    Fetches only the current info and no historical data.
+    On predefined intervals it will fetch the data from coinmarketcap and insert it
+    into a local DB
+  """
   use GenServer, restart: :permanent, shutdown: 5_000
 
   require Sanbase.Utils.Config, as: Config
+  require Logger
 
+  alias Sanbase.DateTimeUtils
   alias Sanbase.Model.{LatestCoinmarketcapData, Project}
   alias Sanbase.Repo
   alias Sanbase.ExternalServices.Coinmarketcap.Ticker
   alias Sanbase.Prices.Store
-
-  # 5 minutes
-  @default_update_interval 1000 * 60 * 5
 
   def start_link(_state) do
     GenServer.start_link(__MODULE__, :ok)
@@ -24,39 +25,57 @@ defmodule Sanbase.ExternalServices.Coinmarketcap.TickerFetcher do
     if Config.get(:sync_enabled, false) do
       Store.create_db()
 
-      GenServer.cast(self(), :sync)
+      Process.send(self(), :sync, [:noconnect])
 
-      update_interval = Config.get(:update_interval, @default_update_interval)
+      update_interval = Config.get(:update_interval) |> String.to_integer()
+
+      Logger.info(
+        "[CMC] Starting TickerFetcher scraper. It will query coinmarketcap every #{
+          update_interval
+        } seconds."
+      )
+
       {:ok, %{update_interval: update_interval}}
     else
       :ignore
     end
   end
 
-  def handle_cast(:sync, %{update_interval: update_interval} = state) do
-    # Fetch current  marketcap
-    tickers = Ticker.fetch_data()
+  def work() do
+    Logger.info("[CMC] Fetching realtime data from coinmarketcap")
+    # Fetch current coinmarketcap data for many tickers
+    {:ok, tickers} = Ticker.fetch_data()
 
-    tickers
-    |> Enum.each(&store_latest_coinmarketcap_data/1)
-
-    tickers
-    |> Enum.flat_map(&Ticker.convert_for_importing/1)
-    |> Store.import()
-
+    # Create a project if it's a new one in the top projects and we don't have it
     tickers
     |> Enum.take(top_projects_to_follow())
     |> Enum.each(&insert_or_create_project/1)
 
-    Process.send_after(self(), {:"$gen_cast", :sync}, update_interval)
+    # Store the data in LatestCoinmarketcapData in postgres
+    tickers
+    |> Enum.each(&store_latest_coinmarketcap_data!/1)
 
-    {:noreply, state}
+    # Store the data in Influxdb
+    tickers
+    |> Enum.map(&Ticker.convert_for_importing/1)
+    |> Store.import()
+
+    Logger.info(
+      "[CMC] Fetching realtime data from coinmarketcap done. The data is imported in the database."
+    )
   end
 
   # Helper functions
 
-  defp store_latest_coinmarketcap_data(ticker) do
-    ticker.id
+  def handle_info(:sync, %{update_interval: update_interval} = state) do
+    work()
+    Process.send_after(self(), :sync, update_interval * 1000)
+
+    {:noreply, state}
+  end
+
+  defp store_latest_coinmarketcap_data!(%Ticker{id: coinmarketcap_id} = ticker) do
+    coinmarketcap_id
     |> LatestCoinmarketcapData.get_or_build()
     |> LatestCoinmarketcapData.changeset(%{
       market_cap_usd: ticker.market_cap_usd,
@@ -71,7 +90,7 @@ defmodule Sanbase.ExternalServices.Coinmarketcap.TickerFetcher do
       percent_change_1h: ticker.percent_change_1h,
       percent_change_24h: ticker.percent_change_24h,
       percent_change_7d: ticker.percent_change_7d,
-      update_time: DateTime.from_unix!(ticker.last_updated)
+      update_time: DateTimeUtils.from_iso8601!(ticker.last_updated)
     })
     |> Repo.insert_or_update!()
   end
