+++ test/sanbase/external_services/coinmarketcap/cryptocurrency_info_test.exs
