+++ test/sanbase/external_services/coinmarketcap/data/coinmarketcap_total_graph_data.json
