+++ test/sanbase/external_services/coinmarketcap/data/cryptocurrency_info.json
