+++ test/sanbase/external_services/coinmarketcap/data/invalid_cryptocurrency_info.json
