+++ test/sanbase/external_services/coinmarketcap/data/pro_cmc_api_2.json
