+++ test/sanbase/external_services/coinmarketcap/logo_fetcher_test.exs
